% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDelta_SKAT0.R
\name{calculateDelta_SKAT0}
\alias{calculateDelta_SKAT0}
\title{Function to calculate delta chi-square scores based on SKAT-O (Step 1 of RIFT)}
\usage{
calculateDelta_SKAT0(y, geno.mat, cov.mat = NULL)
}
\arguments{
\item{y}{outcome vector}

\item{geno.mat}{genotype matrix, requires no missing data}

\item{cov.mat}{covariate matrix, if applicable}
}
\value{
A list object of results
}
\description{
This function takes in y (phenotype), x (genotype matrix) and cov (covariate matrix)
}
\examples{
data(exampleData)
y <- exampleData$PHENOTYPE - 1
geno <- as.matrix(exampleData[,-c(1,2)])
cov <- as.matrix(exampleData$SEX)
results <- calculateDelta_SKAT0(y, geno.mat = geno, cov.mat = cov)
results <- calculateTukeyFences(results)
results <- calculateMAD(results)
results$ind.stats[results$ind.stats$tukey.mild,]
plotDeltaTukey(results)
}
