% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMAD.R
\name{calculateMAD}
\alias{calculateMAD}
\title{Function to calculate MAD and call influential variants}
\usage{
calculateMAD(results, cutoff = 3, constant = 1.4826)
}
\arguments{
\item{results}{output from calculateDelta_SKAT0()}

\item{cutoff}{MAD cutoff for outliers, recommended cutoff of 3}

\item{constant}{MAD constant, recommended constant for normal distribution of 1.4826}
}
\value{
A list object of results
}
\description{
This function takes in the results list object from calculateDelta_SKAT0()
}
\examples{
data(exampleData)
y <- exampleData$PHENOTYPE - 1
geno <- as.matrix(exampleData[,-c(1,2)])
cov <- as.matrix(exampleData$SEX)
results <- calculateDelta_SKAT0(y, geno.mat = geno, cov.mat = cov)
results <- calculateTukeyFences(results)
results <- calculateMAD(results)
results$ind.stats[results$ind.stats$tukey.mild,]
plotDeltaTukey(results)
}
