% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTukeyFences.R
\name{calculateTukeyFences}
\alias{calculateTukeyFences}
\title{Function to calculate Tukey Fences and call influential variants}
\usage{
calculateTukeyFences(results)
}
\arguments{
\item{results}{output from calculateDelta_SKAT0()}
}
\value{
A list object of results
}
\description{
This function takes in the results list object from calculateDelta_SKAT0()
}
\examples{
data(exampleData)
y <- exampleData$PHENOTYPE - 1
geno <- as.matrix(exampleData[,-c(1,2)])
cov <- as.matrix(exampleData$SEX)
results <- calculateDelta_SKAT0(y, geno.mat = geno, cov.mat = cov)
results <- calculateTukeyFences(results)
results <- calculateMAD(results)
results$ind.stats[results$ind.stats$tukey.mild,]
plotDeltaTukey(results)
}
